% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reports.R
\name{get_reports}
\alias{get_reports}
\title{Get WSPD Data}
\usage{
get_reports(start_month = 1, end_month = 12)
}
\arguments{
\item{start_month}{a number representing the month}

\item{end_month}{a number representing the end of the nonth}
}
\description{
This function is used to pull the current incident tracking
    report from online
}
\examples{
\dontrun{

out <- get_reports()

}
}
