% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ward_centers}
\alias{ward_centers}
\title{Ward Centers}
\format{a dataframe with 12 rows and 4 columns:
\describe{
  \item{Ward}{The name of the ward}
  \item{lon}{Type of Call}
  \item{lat}{Where the incident occured}
  \item{geometry}{Date}
   ...
}}
\source{
\url{http://www.cityofws.org}
}
\usage{
ward_centers
}
\description{
The centers of the Wards in Winston-Salem
}
\keyword{datasets}
