% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ward_map}
\alias{ward_map}
\title{Ward Centers}
\format{an sf object}
\source{
\url{http://www.cityofws.org}
}
\usage{
ward_map
}
\description{
The shapefile for the Wards of Winston-Salem
}
\keyword{datasets}
