% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wspd_incidents}
\alias{wspd_incidents}
\title{Winston-Salem Police Incident Logs}
\format{a dataframe with many rows and 6 columns:
\describe{
  \item{case}{Case Number}
  \item{call_type}{Type of Call}
  \item{address}{Where the incident occured}
  \item{incident_date}{Date}
  \item{incident_time}{Time}
  \item{incident_date_time}{Date and time combined}
   ...
}}
\source{
\url{http://www.cityofws.org/crimestats/}
}
\usage{
wspd_incidents
}
\description{
This is a running log of the incident reports posted to the
    Winston-Salem Police Department Website.
}
\keyword{datasets}
